Moved CourierDeliveriesSortView to Order API.
